backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-storage-blob==12.17.0
requests==2.31.0
httpx==0.28.1
pydantic==2.5.0
python-dotenv==1.0.0
six==1.16.0
